requirements.txt
asyncio >= 3.4.3
paho-mqtt >= 1.6.1
tensorflow >= 2.1.0